Decompose cli/_prompts.py (819 lines) into domain-focused submodules
Split _prompts.py into three submodules using a re-export hub pattern:
- _prompts_campaign.py: L3 campaign dispatcher prompt builder
- _prompts_orchestrator.py: L1/L2 cook session prompt + ingredients + greetings
- _prompts_kitchen.py: open-kitchen and fleet-dispatch prompts

_prompts.py reduced to ~50 lines: shared helpers
(_MCP_RETRY_INSTRUCTION, _read_full_sous_chef) + re-exports from submodules.

Ordering is load-bearing: shared helpers must be defined before
re-export imports so that circular imports from submodules
(importing _read_full_sous_chef from _prompts.py) resolve safely.
